Register advancement criterion as early as possible.
