+++ src/main/java/org/spongepowered/common/mixin/core/text/MixinTextComponentBase.java
@@ -52,6 +52,8 @@ import org.spongepowered.common.text.format.SpongeTextColor;
 import java.util.Iterator;
 import java.util.List;
 
+import javax.annotation.Nullable;
+
 @Mixin(TextComponentBase.class)
 public abstract class MixinTextComponentBase implements IMixinTextComponent {
 
@@ -192,10 +194,14 @@ public abstract class MixinTextComponentBase implements IMixinTextComponent {
 
     @Override
     public Text toText() {
-        Text.Builder builder = createBuilder();
+        Text.Builder builder = this.createBuilder();
 
         if (this.style != null) {
             if (this.style.color != null) {
+                final SpongeTextColor spongeColor = SpongeTextColor.of(this.style.color);
+                if (spongeColor != null) {
+                    builder.color(spongeColor);
+                }
                 builder.color(SpongeTextColor.of(this.style.color));
             }
